+++ src/structs/Particle.h
@@ -7,6 +7,6 @@ struct Particle
     Vector2 pos;
     Vector2 inertia;
     Color color;
-    float life;
-    float size;
+    double life;
+    double size;
 };
